fixed move
