+++ pandas/tests/extension/decimal/__init__.py
@@ -1,4 +1,4 @@
-from .array import DecimalArray, DecimalDtype
+from .array import DecimalArray, DecimalDtype, to_decimal, make_data
 
 
 __all__ = ['DecimalArray', 'DecimalDtype', 'to_decimal', 'make_data']
