+++ pandas/tests/extension/json/array.py
@@ -13,6 +13,8 @@ hack around pandas by using UserDicts.
 import collections
 import itertools
 import numbers
+import random
+import string
 import sys
 
 import numpy as np
@@ -179,3 +181,10 @@ class JSONArray(ExtensionArray):
         # cast them to an (N, P) array, instead of an (N,) array of tuples.
         frozen = [()] + [tuple(x.items()) for x in self]
         return np.array(frozen, dtype=object)[1:]
+
+
+def make_data():
+    # TODO: Use a regular dict. See _NDFrameIndexer._setitem_with_indexer
+    return [collections.UserDict([
+        (random.choice(string.ascii_letters), random.randint(0, 100))
+        for _ in range(random.randint(0, 10))]) for _ in range(100)]
