+++ pandas/tests/extension/json/test_json.py
@@ -1,7 +1,5 @@
 import operator
 import collections
-import random
-import string
 
 import pytest
 
@@ -10,16 +8,9 @@ import pandas.util.testing as tm
 from pandas.compat import PY2, PY36
 from pandas.tests.extension import base
 
-from .array import JSONArray, JSONDtype
+from .array import JSONArray, JSONDtype, make_data
 
 pytestmark = pytest.mark.skipif(PY2, reason="Py2 doesn't have a UserDict")
-
-
-def make_data():
-    # TODO: Use a regular dict. See _NDFrameIndexer._setitem_with_indexer
-    return [collections.UserDict([
-        (random.choice(string.ascii_letters), random.randint(0, 100))
-        for _ in range(random.randint(0, 10))]) for _ in range(100)]
 
 
 @pytest.fixture
